Backend/requirements.txt
fastapi 
uvicorn[standard]
python-jose[pyjwt]
python-dotenv
python-multipart
sqlalchemy
passlib